requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.1
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==3.1.3
WTForms==3.2.1
